requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
cffi==2.0.0
discord.py==2.6.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
PyNaCl==1.6.0
python-dotenv==1.1.1
typing_extensions==4.15.0
yarl==1.20.1
yt-dlp==2025.9.5
requests==2.32.5